+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [Unreleased]
 ### Added
 - Add organization value CRUD actions.
+- Add recognition CRUD actions.
 
 ## [0.0.11] - 2017-10-05
 ### Changed